SAF-102: Static: Refactor sidebar files
As per atomic design, refactored all sidebar-related files. First,
SidebarTemplate.tsx was broken down into two atoms - one for the user
name and the other for the user picture. Second, Sidebar.tsx was
modified to use props to populate these atoms.
